cleanup/improvements

+++ src/main/java/dev/zwazel/internal/game/tank/TankConfig.java
@@ -1,4 +1,4 @@
 package dev.zwazel.internal.game.tank;
 
-public record TankConfig(long speed) {
+public record TankConfig(float speed) {
 }
